study - mysite06 guestbook-ajax

+++ mysite06/src/main/java/com/poscodx/mysite/vo/GuestbookVo.java
@@ -1,15 +1,15 @@
 package com.poscodx.mysite.vo;
 
 public class GuestbookVo {
-	private Long no;
+	private int no;
 	private String name;
 	private String contents;
 	private String password;
 	private String regDate;
-	public Long getNo() {
+	public int getNo() {
 		return no;
 	}
-	public void setNo(Long no) {
+	public void setNo(int no) {
 		this.no = no;
 	}
 	public String getName() {
